Add files via upload
This is a Porodomo timer project where you complete 25 min of focus on your activity of choice, once you've completed your 25 min time cycle and Waifu will be generated for you. However, if you leave the page before the timer is up she will pass and go to the Waifu graveyard.